+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 This list is not currently intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate:
 
-### v3.179.2
+### v3.180.0 (uncut)
+- [all] Update various dependencies.
+- [http4k-testing-hamcrest] Improve messages of Hamkrest matchers. H/T @albertlatacz
+- [http4k-cloudnative] Fix #291 - Readiness check result when there are > 2 checks may not report the correct result. H/T @alfi
 - [http4-security-oauth] [Possibly breaking] Making client_secret optional in AuthorizationCodeAccessTokenRequest to support non client_secret flows.
 
 ### v3.179.1